# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Airbnb Clone API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/bookings/quote": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Quote Booking",
        "description": "Calculate price breakdown for given listing + dates.\nReturns 404 if listing not found, 409 if dates unavailable.",
        "operationId": "quote_booking_api_bookings_quote_get",
        "parameters": [
          {
            "name": "listing_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Listing Id"
            }
          },
          {
            "name": "check_in",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check In"
            }
          },
          {
            "name": "check_out",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Check Out"
            }
          },
          {
            "name": "adults",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Adults"
            }
          },
          {
            "name": "children",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Children"
            }
          },
          {
            "name": "infants",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Infants"
            }
          },
          {
            "name": "pets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Pets"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Create Booking Endpoint",
        "description": "Create a confirmed booking (authoritative pricing, overlap-safe transaction).",
        "operationId": "create_booking_endpoint_api_bookings_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "List Bookings",
        "description": "Get current user's bookings (guest view).",
        "operationId": "list_bookings_api_bookings_get",
        "parameters": [
          {
            "name": "tab",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(all|upcoming|past)$",
              "default": "all",
              "title": "Tab"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingResponse"
                  },
                  "title": "Response List Bookings Api Bookings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{booking_id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Booking",
        "description": "Get single booking (guest or host of that listing).",
        "operationId": "get_booking_api_bookings__booking_id__get",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{booking_id}/cancel-preview": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Preview Cancel",
        "description": "Preview refund before cancelling.",
        "operationId": "preview_cancel_api_bookings__booking_id__cancel_preview_get",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{booking_id}/cancel": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Cancel Booking Endpoint",
        "description": "Cancel a booking and process refund.",
        "operationId": "cancel_booking_endpoint_api_bookings__booking_id__cancel_post",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{booking_id}/review": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Create Review Endpoint",
        "description": "Submit a review for a completed stay.",
        "operationId": "create_review_endpoint_api_bookings__booking_id__review_post",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "listing_title": {
            "type": "string",
            "title": "Listing Title"
          },
          "listing_city": {
            "type": "string",
            "title": "Listing City"
          },
          "listing_country": {
            "type": "string",
            "title": "Listing Country"
          },
          "listing_photo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Photo"
          },
          "check_in": {
            "type": "string",
            "title": "Check In"
          },
          "check_out": {
            "type": "string",
            "title": "Check Out"
          },
          "adults": {
            "type": "integer",
            "title": "Adults"
          },
          "children": {
            "type": "integer",
            "title": "Children"
          },
          "infants": {
            "type": "integer",
            "title": "Infants"
          },
          "pets": {
            "type": "integer",
            "title": "Pets"
          },
          "nights": {
            "type": "integer",
            "title": "Nights"
          },
          "price_per_night": {
            "type": "integer",
            "title": "Price Per Night"
          },
          "nightly_total": {
            "type": "integer",
            "title": "Nightly Total"
          },
          "cleaning_fee": {
            "type": "integer",
            "title": "Cleaning Fee"
          },
          "service_fee": {
            "type": "integer",
            "title": "Service Fee"
          },
          "total_price": {
            "type": "integer",
            "title": "Total Price"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "payment_status": {
            "type": "string",
            "title": "Payment Status"
          },
          "refund_amount": {
            "type": "integer",
            "title": "Refund Amount"
          },
          "confirmation_code": {
            "type": "string",
            "title": "Confirmation Code"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "can_cancel": {
            "type": "boolean",
            "title": "Can Cancel"
          },
          "can_review": {
            "type": "boolean",
            "title": "Can Review"
          },
          "has_review": {
            "type": "boolean",
            "title": "Has Review"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "listing_id",
          "listing_title",
          "listing_city",
          "listing_country",
          "listing_photo",
          "check_in",
          "check_out",
          "adults",
          "children",
          "infants",
          "pets",
          "nights",
          "price_per_night",
          "nightly_total",
          "cleaning_fee",
          "service_fee",
          "total_price",
          "status",
          "payment_status",
          "refund_amount",
          "confirmation_code",
          "phase",
          "can_cancel",
          "can_review",
          "has_review",
          "created_at",
          "cancelled_at"
        ],
        "title": "BookingResponse"
      },
      "CancelPreviewResponse": {
        "properties": {
          "booking_id": {
            "type": "string",
            "title": "Booking Id"
          },
          "refund_amount": {
            "type": "integer",
            "title": "Refund Amount"
          },
          "refund_policy": {
            "type": "string",
            "title": "Refund Policy"
          },
          "nights_elapsed": {
            "type": "integer",
            "title": "Nights Elapsed"
          },
          "nights_total": {
            "type": "integer",
            "title": "Nights Total"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "refund_amount",
          "refund_policy",
          "nights_elapsed",
          "nights_total"
        ],
        "title": "CancelPreviewResponse"
      },
      "CreateBookingRequest": {
        "properties": {
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "check_in": {
            "type": "string",
            "title": "Check In"
          },
          "check_out": {
            "type": "string",
            "title": "Check Out"
          },
          "adults": {
            "type": "integer",
            "title": "Adults"
          },
          "children": {
            "type": "integer",
            "title": "Children",
            "default": 0
          },
          "infants": {
            "type": "integer",
            "title": "Infants",
            "default": 0
          },
          "pets": {
            "type": "integer",
            "title": "Pets",
            "default": 0
          },
          "card_token": {
            "type": "string",
            "title": "Card Token"
          }
        },
        "type": "object",
        "required": [
          "listing_id",
          "check_in",
          "check_out",
          "adults",
          "card_token"
        ],
        "title": "CreateBookingRequest"
      },
...
      "CreateReviewRequest": {
        "properties": {
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "rating",
          "comment"
        ],
        "title": "CreateReviewRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuoteResponse": {
        "properties": {
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "check_in": {
            "type": "string",
            "title": "Check In"
          },
          "check_out": {
            "type": "string",
            "title": "Check Out"
          },
          "nights": {
            "type": "integer",
            "title": "Nights"
          },
          "price_per_night": {
            "type": "integer",
            "title": "Price Per Night"
          },
          "nightly_total": {
            "type": "integer",
            "title": "Nightly Total"
          },
          "cleaning_fee": {
            "type": "integer",
            "title": "Cleaning Fee"
          },
          "service_fee": {
            "type": "integer",
            "title": "Service Fee"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "USD"
          }
        },
        "type": "object",
        "required": [
          "listing_id",
          "check_in",
          "check_out",
          "nights",
          "price_per_night",
          "nightly_total",
          "cleaning_fee",
          "service_fee",
          "total"
        ],
        "title": "QuoteResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}